requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pytesseract>=0.3.10
Pillow>=9.0.0

# For Premiere Pro label automation (premiere_label_sender.py)
pyautogui>=0.9.54
pygetwindow>=0.0.9
